==> 07b_Figures_2024_labelmap.Rmd <==
---
title: "07b map for methods chapter - 2024 "
output: 
  html_document:
    keep_md: true
    toc: true
    toc_float: true
editor_options: 
  chunk_output_type: console
---

Overview maps (types of stations) for Methods chapter (section 3)  
* using UTM zone 32V as projection  

## Overview of process:    
1. Packages and load scripts      
2. Data, includes    
    - station data  
    - background map (coast- + county lines)  
    - rivers  
    - the last two: downloaded as shape (UTM33), but coordinares extracted and converted to UTM32  
3. Defines inset map boxes using a leaflet map; therefore boxes are defined using lat/long  
    - boxes are saved as a list, each of which defining bottom left and top right corner   
4. Prepare data sets in UTM32  
    - station data   
    - box_corners (used to make the inset plots, 'make_subplot_nolabels', and to filter stations for these)
    - box_polygons (used to plot inset rectangles on the big map, 6b)  
5. Is used only as a fast way to make a draft map with labels (used as help for 6a)
6. Build gg3a plot
    - Build gg2_points plot based on background map (gg_back + rivers) + labels, group by group  
    - gg3a map = gg2_points map plus boxes   
7. Functions for making insets  
8. Make insets including labels. the resukt is a list of ggplot objects)
9. Make final plot by adding legend (gg3b), then plot gg3b plus each of the insets   
    - position of insets is quite a bit try and error   

## Overview of some important objects and where they are created      

data:  
- 2a: df_stations_all (read)
- 2d: df_stations_south, df_stations_north  

map data:
- 3a: map_counties_longlat - geojson file from Geonorge  
- 3b: nc_rivers_full - geojson file from Geonorge  
- 3c: nc_lakes_all - GML file from Geonorge  


## 1. Prerequisites  

### a. Whether and where to save figures  
```{r}
#
# Save figures to file, or just show them in this report?
#
# NOTE: we now use a separate R file (see 9b in this script) 
# - `07b_Figures_2020_labelmap_combine.R`  
#
# save_plots <- TRUE
save_plots <- FALSE

#
# Year
#
year <- 2024

#
# Where to save figures
#
folder_fig <- glue::glue("Figures/{year}_main")

if (!dir.exists(folder_fig))
  dir.create(folder_fig)

```

### b. Packages + map  
```{r}

# library(tidyverse)  # filter, group_by, summary, etc.
library(dplyr)  # filter, group_by, summary, etc.
library(readr)  # write_csv()
library(ggeasy)
library(gridGraphics)
# library(maps)       # Norway map (NOTE: must be loaded after tidyverse)
# library(mapdata)
library(readxl)
library(glue)         # for glue() function  

# map_norway <- map_data("world", "Norway")
# map_norway_h <- readRDS("Data_input/map_norway_hires.RData")

# Long/lat map (including the smaller islands) - lacking Mjøsa, though! 
#   (and including some North Finnish lakes?)
load("Data_input/Norway_coastline_longlat2.RData")
# rename
map_norway_h <- norway_coast_longlat2
rm(norway_coast_longlat)

# Test plot
# ggplot(map_norway_h) + geom_path(aes(long, lat, group = group))

library(mapproj)    # mapproject
library(ggrepel)    # geom_text_repel()
library(sp)         # SpatialPoints(), CRS(), spTransform()
library(sf)         # st_read
library(ggimage)    # geom_subview(), theme_transparent()

library(leaflet)

library(cowplot)

library(scatterD3)  # used to position labels  

source("07_Figures_functions.R")

# Define crs strings for conversion between long/lat and UTM
crs_longlat <- "+proj=longlat +ellps=WGS84 +datum=WGS84"
crs_utm <- "+proj=utm +zone=32 +ellps=WGS84 +datum=WGS84 +units=m"  # note: zone 32

```


## 2. Data

### a. Data   

* From excel  

```{r, warning=FALSE}

# df_stations_all_OLD <- readRDS("Data/2021/07a_df_stations_2020.rds")
# str(df_stations_all_OLD)

df_stations_all <- readRDS(glue("Data/{year}/07a_df_stations_{year}.rds"))


if (FALSE){
  
  # For reading from excel 
  coltypes <- c("text", "numeric", "numeric", "numeric", "text")
  df_stations_all <- read_excel("Data_input/2022/KartOgFigurgrunnlag2021.xlsx", 
                                range = "A1:E34", 
                                col_types = coltypes) %>%
    rename(Shortname = Rapportnavn,
           Lat = Biologi_Lat,
           Long = Biologi_Long)

  str(df_stations_all)
  
}



```

### b. Stations, add UTM coordinates
```{r}

# The default transformation for 'add_transformed_coord' is from latlong to 
df_stations_all <- add_transformed_coord(df_stations_all, 
                                     name_x = "Long", name_y = "Lat",
                                     from = "longlat", to = "utm", to_zone = 32)  

write_csv(df_stations_all %>% select(-Stasjonstype),
          glue("Data_input/{year}/KartOgFigurgrunnlag{year}.csv"))  

# write_csv(very_simple_map %>% add_transformed_coord(),
#           "Data_input/2022/very_simple_map.csv")

```

### c. Plot all stations  
```{r}

df_stations_all$Stasjonstype %>% table()

very_simple_map <- map_data("world", "Norway")

ggplot(df_stations_all, aes(Long, Lat, color = Stasjonstype)) +
  annotation_map(very_simple_map, fill = "lightgreen") +
  geom_point() +
  # geom_text_repel() +
  coord_map("lambert", 15, 65)  


```

### d. Split into North and South parts      

* NOTE: how the stations/maps are split must be done depending on the stations used for each year  

```{r}

df_stations_south <- df_stations_all %>%
  filter(Lat < 61)

df_stations_north <- df_stations_all %>%
  filter(Lat >= 61)

if (TRUE){
  
  ggplot(df_stations_south, aes(Long, Lat, color = Stasjonstype)) +
    annotation_map(very_simple_map, fill = "lightgreen") +
    geom_point() +
    # geom_text_repel() +
    coord_map("lambert", 15, 65) 
  
  ggplot(df_stations_north, aes(Long, Lat, color = Stasjonstype)) +
    annotation_map(very_simple_map, fill = "lightgreen") +
    geom_point() +
    # geom_text_repel() +
    coord_map("lambert", 15, 65) 
  
}

```
## 3. Map data  

### a. County map (fylke)   

```{r}
# 
# Gotten from Geonorge ("Norske fylker og kommuner illustrasjonsdata 2019 (klippet etter kyst)")
#

map_counties_longlat <- sf::st_read(
  "../../Mapdata/Basisdata_0000_Norge_25833_NorskeFylkerKommunerIllustrasjonsdata2024_GeoJSON/GeoJSON/Fylker_simple300.geojson") 
plot(map_counties_longlat["navn"])
map_counties_longlat[["navn"]]

# Transform to UTM32
map_counties <- map_counties_longlat %>% 
  sf::st_transform(25832)

```

### b. Rivers  

```{r}

# NEW (geojson)
nc_rivers_full <- st_read(
  "../../Mapdata/NVE_ELVIS_elvenett/NVEData/Elv_Hovedelv.geojson")
str(nc_rivers_full)
table(addNA(nc_rivers_full$nivaa))

# test
nc_rivers_full %>%
  filter(nivaa %in% c("1a")) %>% 
  .["elvenavn"] %>% 
  plot()

nc_rivers_nivaa2_full <- nc_rivers_full %>%
  filter(nivaa %in% c("1a", "2a"))

# plot(nc_rivers_nivaa2_full["elvenavn"])

# Transform to UTM32
nc_rivers <- nc_rivers_nivaa2_full %>% 
  sf::st_transform(25832)

```

### c. Lakes  

```{r}

fn <- "../../Mapdata/Basisdata_0000_Norge_25833_N1000Kartdata_GML/Basisdata_0000_Norge_25833_N1000Arealdekke_GML.gml"
st_layers(fn)
nc_lakes_all <- st_read(fn, layer = "Innsjø")
nc_lakes_all$area <- st_area(nc_lakes_all)
range(nc_lakes_all$area)
# hist(nc_lakes_all$area, breaks = seq(0,400,5)*1E6)

# Keep only the 50% largest lakes
min <- quantile(nc_lakes_all$area, 0.5)
nc_lakes <- nc_lakes_all %>%
  filter(area > min) %>% 
  # Transform to UTM32
  sf::st_transform(25832)

```

### d. Stations  
```{r}

# see example in 'st_as_sf'  
sf_stations_so = st_as_sf(df_stations_south, coords = c("x", "y"), crs = 25832)
sf_stations_no = st_as_sf(df_stations_north, coords = c("x", "y"), crs = 25832)

```


## 4. Crop layers 

### a. Use leaflet to find corners of maps  

```{r}

if (FALSE){
  
  # advce: do this in 
  
  library(leaflet)
  library(leafem)    # for addMouseCoordinates
  
  df_counties_longlat <- map_counties %>% 
    st_coordinates() %>% 
    as_tibble()
  
  leaflet(df_stations_south) %>% 
    # leaflet(df_stations_north) %>% 
    addTiles() %>% 
    addMarkers(lng = ~Long, lat = ~Lat) %>% 
    leafem::addMouseCoordinates()
  
}

# Southern part:
# 9.35, 59.52, 11.41, 61.03
# Northern part
# 8.65, 62.60, 16.70, 67.53:

```

### b. Crop 

```{r}

# explicity, tell R that attribute variables are constant throughout the geometry:
# See https://github.com/r-spatial/sf/issues/406
st_agr(map_counties) <- "constant"
st_agr(nc_rivers) <- "constant"
st_agr(nc_lakes) <- "constant"
# (st_crop still works if we don't do this, but will give the warning message
# "attribute variables are assumed to be spatially constant throughout all geometries".)

map_counties_so <- st_crop_from_latlong(map_counties, 9.35, 59.52, 11.41, 61.03)
nc_rivers_so <- st_crop_from_latlong(nc_rivers, 9.35, 59.52, 11.41, 61.03)
nc_lakes_so <- st_crop_from_latlong(nc_lakes, 9.35, 59.52, 11.41, 61.03)

map_counties_no <- st_crop_from_latlong(map_counties, 8.65, 62.60, 16.70, 67.53)
nc_rivers_no <- st_crop_from_latlong(nc_rivers, 8.65, 62.60, 16.70, 67.53)
nc_lakes_no <- st_crop_from_latlong(nc_lakes, 8.65, 62.60, 16.70, 67.53)

```

## 5. Make overlay plots  

* Using ggrepel for temporay placement of labels  

### a. Common settings  
```{r}

landcol <- "navajowhite"
bordercol <- "grey30"
rivercol <- "lightskyblue"
textcol <- "#0000FF"        # corresponds to a color in the Inkscape color menu
stationtype_palette <- "Set1"
stationtype_shape <- c(21,22,24)

```


### b. Southern part 

```{r}

# for labels, see example in the ggrepel documentation
# https://ggrepel.slowkow.com/articles/examples#label-sf-objects

gg_so_background <- ggplot() +
  geom_sf(data = map_counties_so, fill = landcol, color = bordercol) +
  geom_sf(data = nc_rivers_so, color = rivercol) +
  geom_sf(data = nc_lakes_so, color = rivercol) +
  theme_minimal() +
  theme(axis.title = element_blank(),
        axis.text = element_blank(),
        axis.ticks = element_blank(),
  )

gg_so <- gg_so_background +
  scale_fill_brewer(palette = stationtype_palette) +
  scale_shape_manual(values = stationtype_shape) +
  ggrepel::geom_text_repel(
    data = sf_stations_so,
    aes(label = Shortname, geometry = geometry),
    stat = "sf_coordinates",
    color = textcol, 
    hjust = 0, force = 1, force_pull = 0.4, nudge_x = 5000, 
    # box.padding = 0.5, point.padding = 0.5,
    min.segment.length = 0.1
  )

gg_so

```


### c. Northern  part 

```{r}

# as southern part, except: 
#   nudge_x increased

gg_no_background <- ggplot() +
  geom_sf(data = map_counties_no, fill = landcol, color = bordercol) +
  geom_sf(data = nc_rivers_no, color = rivercol) +
  geom_sf(data = nc_lakes_no, color = rivercol) +
  theme_minimal() +
  theme(axis.title = element_blank(),
        axis.text = element_blank(),
        axis.ticks = element_blank(),
  )

gg_no <- gg_no_background +
  geom_sf(data = sf_stations_no, aes(fill = Stasjonstype, shape = Stasjonstype)) +
  scale_fill_brewer(palette = stationtype_palette) +
  scale_shape_manual(values = stationtype_shape) +
  ggrepel::geom_text_repel(
    data = sf_stations_no,
    aes(label = Shortname, geometry = geometry),
    stat = "sf_coordinates",
    color = textcol, 
    hjust = 0, force = 1, force_pull = 0.4, nudge_x = 10000, 
    # box.padding = 0.5, point.padding = 0.5,
    min.segment.length = 0.1
  )

gg_no

```



## 6. Save maps  

* Post-processing in Inkscape (moving labels and making some new lines)  
    - See Readme file for some tips  

```{r}

# Save background maps as ggplot objects
saveRDS(gg_so_background, glue("{folder_fig}/07b_so_background.rds"))
saveRDS(gg_no_background, glue("{folder_fig}/07b_no_background.rds"))

# Save plots in svg format (editable) 
ggsave(glue("{folder_fig}/07b_labelmap_so.svg"), gg_so + easy_remove_legend(), width = 7, height = 7)
ggsave(glue("{folder_fig}/07b_labelmap_no.svg"), gg_no + easy_remove_legend(), width = 7, height = 12)
ggsave(glue("{folder_fig}/07b_labelmap_so_with_legend.svg"), gg_so, width = 7, height = 7)

```

## 7. Convert to png  

* After post-processing in Inkscape  

```{r}

library(magick)

for (area in c("so","no")){
  fn_svg <- glue("{folder_fig}/07b_labelmap_{area}_adj.svg")
  dir(folder_fig)
  if (file.exists(fn_svg)){
    my_image <- image_read(fn_svg, density = 200)
    image_info(my_image)
    my_png <- image_convert(my_image, format="png")
    image_write(my_png, sub(".svg$", ".png", fn_svg))
    message("File ", sQuote(fn_svg), " converted to png")
  } else {
    warning("File ", sQuote(fn_svg), " not found")
  }
}

```
