==> 07c_Figures_2024_qualitymap.Rmd <==
---
title: "07b map for results chapter - 2024"
author: "DHJ"
date: "May 2025"
output: html_document
---


"Quality maps" showing one pie per station for Results chapter     
* Pie: 
    - color of left half shows ecological quality (five colours)
    - colour of right half shows chemical quality  (blue or red)  
  
## Code overview   
1. Prerequisites  
    - important: set *year*
    - load packages, load functions from script
2. loads station data, split stations by postiion (typically latitude)  
3. Makes pie ggplot objects using the functions make_pie() and make_pie_from_colours()   
    - These are stored in the column 'pie' in th data frame 'df_pies'   
4. Makes map (only the circle symbols, labels are addedusing Inkscape later)   
    - Load background maps (ggplot objects) made in script 07b  
    - Split pies into North and South parts  
    - Using ggplot, plot background maps and then 'df_pies'    

## 1. Prerequisites  

### a. Whether and where to save figures  
```{r}
#
# Year
#
year <- 2024

#
# Where to save figures
#
folder_fig <- glue::glue("Figures/{year}_main")

```

### b. Libs + map  
```{r}

library(dplyr)  # filter, group_by, summary, etc.
library(gridGraphics)
library(maps)       # Norway map (NOTE: must be loaded after tidyverse)
library(readxl)

# map_norway <- map_data("world", "Norway")
# map_norway_h <- readRDS("Data_input/map_norway_hires.RData")

# Long/lat map (including the smaller islands) - lacking Mjøsa, though! 
#   (and including some North Finnish lakes?)
load("Data_input/Norway_coastline_longlat2.RData")
# rename
map_norway_h <- norway_coast_longlat2
rm(norway_coast_longlat)

# Test plot
# ggplot(map_norway_h) + geom_path(aes(long, lat, group = group))

# library(mapproj)    # mapproject
library(ggrepel)    # geom_text_repel()
library(sp)         # SpatialPoints(), CRS(), spTransform()
library(sf)         # st_read
library(ggimage)    # geom_subview(), theme_transparent()

library(glue)       # used in ggplot2() function
library(leaflet)

library(cowplot)

source("07_Figures_functions.R")

```
### c. Define colours for classes  
```{r, fig.width=3.5, fig.height=2.5}

# Order in plots etc.
class_levels <- c("Svært dårlig", "Dårlig", "Moderat", "God", "Svært god", "Uklassifisert")

# Read standard colors
df_colors <- read_excel("Data_input/Fargekoder RGB for tilstandsklassifisering.xlsx", sheet = 2) %>%
  mutate(
    Status_no = case_when(
      Status %in% "High" ~ "Svært god",
      Status %in% "Good" ~ "God",
      Status %in% "Moderate" ~ "Moderat",
      Status %in% "Poor" ~ "Dårlig",
      Status %in% "Bad" ~ "Svært dårlig",
      Status %in% "Unclassified" ~ "Uklassifisert"),
    Status_no = factor(Status_no, levels = class_levels),
    Class_color = rgb(R/255, G/255, B/255)
  ) %>%
  arrange(Status_no)
df_colors


# Show colors
par(mar = c(0,0,2,0))
pie(rep(1,6), 
    col = df_colors$Class_color, 
    labels = df_colors$Status_no, 
    main = "Colors for nEQr classes")  

# Also make named list
class_colors <- df_colors$Class_color
names(class_colors) <- df_colors$Status_no



```

## 2. Data   

### a. Data from script 7a and 7b  
```{r}

# From script 7a
df_index_barplot <- readRDS(glue("Data/{year}/07a_df_index_barplot_{year}.rds"))
df_stations_all <- readRDS(glue("Data/{year}/07a_df_stations_{year}.rds"))

df_stations_all <- add_transformed_coord(df_stations_all, 
                                     name_x = "Long", name_y = "Lat",
                                     from = "longlat", to = "utm", to_zone = 32)  

```


### b. Plot all stations  

```{r}

ggplot(df_stations_all, aes(Long, Lat)) +
  geom_point() +
  coord_map("lambert", 15, 65)  

```


### c. Split into North and South Norway    

```{r}

df_stations_south <- df_stations_all %>%
  filter(Lat < 61)

df_stations_north <- df_stations_all %>%
  filter(Lat >= 61)

df_stations_all %>%
  mutate(region = ifelse(Lat >= 61, "North", "South")) %>%
  ggplot(aes(Long, Lat, color = Stasjonstype)) +
  annotation_map(map_norway_h, fill = "lightgreen") +
  geom_point() +
  facet_wrap(vars(region))    # , scales = "free"


```




## 3. Make pies  

### a. Start making 'df_pies'  
Also writes df_stations_vf to excel (later manipulated semi-manually and used in part 6)    
- Note that we use 'df_stations_all', i.e. also include the Northern ones   
```{r}

# Start with location data on Vannforekomst-level
df_pies_all <- df_stations_all %>%
  left_join(df_index_barplot %>% select(Rapportnavn, nEQR_ecol_klasse),
            by = c("Shortname" = "Rapportnavn")) %>%
  rename(Class_ecol = nEQR_ecol_klasse)


```


### b. Ecological class, levels and colors    
```{r}


# df_colors$Status_no
# class_levels <- c("Svært dårlig", "Dårlig", "Moderat", "God", "Svært god", "Uklassifisert")


df_pies_all <- df_pies_all %>%
  left_join(
    df_colors %>% select(Status_no, Class_color),
    by = c("Class_ecol" = "Status_no")
  ) %>%
  rename(Colour_ecol = Class_color)  

ggplot(df_pies_all, aes(Class_ecol, Index_ecol, fill = Class_ecol)) + 
  geom_point(shape = 21, size = 2) +
  scale_fill_manual(values = class_colors, drop = FALSE) +
  theme(panel.background = element_rect(fill = "grey70"), panel.grid = element_blank())

```


### c. Chemical class, levels and colors    

* Good chemical status is denoted as "MO" this year

```{r}

table(addNA(df_pies_all$Class_chem))

df_pies_all <- df_pies_all %>%
  mutate(
    Colour_chem = case_when(
      is.na(Class_chem) ~ class_colors["Uklassifisert"],
      Class_chem %in% c("God", "MO") ~ class_colors["Svært god"],
      Class_chem %in% "Ikke god" ~ class_colors["Svært dårlig"],
      Class_chem %in% c("i.d.", "ID") ~ class_colors["Uklassifisert"])
  )

xtabs(~addNA(Colour_chem) + addNA(Class_chem), df_pies_all)

```


### d. Make pies from data frame 'df_pies_all', and add to 'df_pies_all'  

```{r, fig.width=3, fig.height=3}      

# Function - make pie for line number i
make_pie_linenumber_indexvalues <- function(i){
  df <- df_pies_all[i, c("Colour_ecol", "Colour_chem")]
  make_pie_from_colours(as.character(df))
}

# TEST
make_pie_linenumber_indexvalues(1)
make_pie_linenumber_indexvalues(2)
make_pie_linenumber_indexvalues(3)

# test - should make 3 plots with one pie in each  
if (FALSE){
  grep("03.ORK1", df_pies_all$Shortname) %>% make_pie_linenumber_indexvalues()
  grep("02.ORK2", df_pies_all$Shortname) %>% make_pie_linenumber_indexvalues()
  grep("01.ORK3", df_pies_all$Shortname) %>% make_pie_linenumber_indexvalues()
}

# Make all pies (a list)
pies <- 1:nrow(df_pies_all) %>% purrr::map(~make_pie_linenumber_indexvalues(.))
length(pies)
# check:
# pies[[1]]

# Add pies to dataframe (as a list column)
df_pies_all$pie = pies
# check:
# df_pies_all$pie[1]

# df <- df_pies_all %>% filter(!Shortname %in% c(group1, group2a, group2b))
# df$pie[1]

```

### e. Test  
```{r}

df_pies_test <- df_pies_all[c(1,5,10),]

#
# Width and height of pies
#   
df_pies_test$width = 355E3    # Sets size of pies
df_pies_test$height = 35E3   

df_stations_all %>%
  ggplot(aes(x, y, color = Stasjonstype)) +
  geom_point() +
  geom_point(data = df_pies_test, aes(x = x, y = y)) +
  geom_subview(data = df_pies_test, 
               aes(x = x, y = y, subview = pie, width = width, height = height)
               ) +
  coord_fixed()


```

## 4. Map   

* We didn't have to make any boxes this time, so this is the entire map     

### a. Load background maps  
```{r}

# Load background maps as ggplot objects
gg_so_background <- readRDS(glue("{folder_fig}/07b_so_background.rds"))
gg_no_background <- readRDS(glue("{folder_fig}/07b_no_background.rds"))

```

### b. Split pies into North and South parts      

* NOTE: use the same split as used in script 07b for the same year (part 2d "Split into North and South parts")  
    - how the stations/maps differs between years  

```{r}

df_pies_so <- df_pies_all %>% 
  filter(Lat < 61)

df_pies_no <- df_pies_all %>% 
  filter(Lat >= 61)

```

### c. South  

```{r}

df_pies_so$width = 5E3    # Sets size of pies
df_pies_so$height = 5E3   

gg_so <- gg_so_background +
  geom_subview(data = df_pies_so, 
               aes(x = x, y = y, subview = pie, width = width, height = height)
               )

gg_so  

```
### d. North  

```{r}

df_pies_no$width = 9E3    # Sets size of pies
df_pies_no$height = 9E3   

gg_no <- gg_no_background +
  geom_point(data = df_pies_no, aes(x = x, y = y), pch = 20) +
  geom_subview(data = df_pies_no, 
               aes(x = x, y = y, subview = pie, width = width, height = height)
               )

gg_no  

```

### e. Save plots  

* We add labels later   
    - open the saved plot in Inkscape  
    - also open the adjusted version ('adj' in file name) of the corresponding 07b plot ('so' or 'no') from script
        - i.e. "07b_labelmap_so_adj.svg" or "07b_labelmap_no_adj.svg"
    - copy all labels from this map (Edit > Select Same > Object type) and paste into the 'qualitymap'  
    - copy all lead lines from this map (Edit > Select Same > Stroke color) to the 'qualitymap'  
    - you may want to adjust the positions a bit  
    - saved with "adj" in file name  

```{r}

# Save plots in svg format (editable) 
ggsave(glue("{folder_fig}/07c_qualitymap_so.svg"), gg_so, width = 7, height = 7)
ggsave(glue("{folder_fig}/07c_qualitymap_no.svg"), gg_no, width = 7, height = 12)


```


## 5. Convert to png  

* After post-processing in Inkscape  

```{r}

library(magick)

for (area in c("so","no")){
  fn_svg <- glue("{folder_fig}/07c_qualitymap_{area}_adj.svg")
  dir(folder_fig)
  if (file.exists(fn_svg)){
    my_image <- image_read(fn_svg, density = 200)
    image_info(my_image)
    my_png <- image_convert(my_image, format="png")
    image_write(my_png, sub(".svg$", ".png", fn_svg))
    message("File ", sQuote(fn_svg), " converted to png")
  } else {
    warning("File ", sQuote(fn_svg), " not found")
  }
}

```
