==> 07d_Figures_2024_barplots.Rmd <==
---
title: "Barplot figures (2024 data)"
author: "DHJ"
date: "May 2025"
output: 
  html_document:
    keep_md: true
    toc: true
    
---

# Bar plots for nEQR quality   

Plots for data 2024 (report made in 2025)  
* Data are prepared in script 07a   
* Plotting uses function `make_barplot_index()` in script 07..  

## 1. Prerequisites  

### a. Whether and where to save figures  
```{r}
#
# Save figures to file, or just show them in this report?
#
#
save_plots <- TRUE
# save_plots <- FALSE

#
# Year
#
year <- 2024

#
# Where to save figures
#
folder_fig <- glue("Figures/{year}_main")

# dir.exists(folder_fig)
# dir(folder_fig)

```

### b. Libs + map  
```{r, results='hide', warning=FALSE, message=FALSE}

library(tidyverse)  # filter, group_by, summary, etc.
library(readxl)
library(cowplot)
library(glue)

source("07_Figures_functions.R")

# Define crs strings for conversion between long/lat and UTM
crs_longlat <- "+proj=longlat +ellps=WGS84 +datum=WGS84"
crs_utm <- "+proj=utm +zone=32 +ellps=WGS84 +datum=WGS84 +units=m"

source("07_Figures_functions.R")

```

### c. Define colours for classes  
```{r, fig.width=3.5, fig.height=2.5}

# Order in plots etc.
class_levels <- c("Svært dårlig", "Dårlig", "Moderat", "God", "Svært god", "Uklassifisert")

# Read standard colors
df_colors <- read_excel("Data_input/Fargekoder RGB for tilstandsklassifisering.xlsx", sheet = 2) %>%
  mutate(
    Status_no = case_when(
      Status %in% "High" ~ "Svært god",
      Status %in% "Good" ~ "God",
      Status %in% "Moderate" ~ "Moderat",
      Status %in% "Poor" ~ "Dårlig",
      Status %in% "Bad" ~ "Svært dårlig",
      Status %in% "Unclassified" ~ "Uklassifisert"),
    Status_no = factor(Status_no, levels = class_levels),
    Class_color = rgb(R/255, G/255, B/255)
  ) %>%
  arrange(Status_no)
# df_colors


# Show colors
par(mar = c(0,0,2,0))
pie(rep(1,6), 
    col = df_colors$Class_color, 
    labels = df_colors$Status_no, 
    main = "Colors for nEQr classes")  

# Also make named list
class_colors <- df_colors$Class_color
names(class_colors) <- df_colors$Status_no

```

## 2. Data   

### a. Data from script 7a 
```{r}

# From script 7a
# df_index <- readRDS("Data/2022/07a_df_index_2019.rds")
df_index_barplot <- readRDS(glue("Data/{year}/07a_df_index_barplot_{year}.rds") # %>%
#  rename(Tilstandsklasse = nEQR_ecol_klasse)

# From script 7b, part 4c
# df_stations <- readRDS("Data/2022/07b_df_stations.rds")   # with UTM 32 coordinates

```


## 3. Bar plots

### a. Eutrofiering
* no figure this year  

### b. Forsuring (acidification)  
* no figure this year  

### c. Samlet økologisk tilstand 
```{r, fig.height=8.5}

label <- "Samlet økologisk tilstand (nEQR)"
plotno <- "1c"

gg1 <- make_barplot_index(df_index_barplot, "nEQR_ecol", label)
gg2 <- gg1 + 
  # Add all-white label (to cover the vertical lines)
  # (this is a workaround, as 'colour' seems to control both text and border colour...)
  # geom_label(aes(Rapportnavn, Indeksverdi, label = Label),
  #            hjust = 0, nudge_y = 0.005, nudge_x = 0, color = "white", fill = "white") +
  # Add text
  geom_text(aes(Rapportnavn, Indeksverdi, label = Rapportnavn),
             hjust = 0, nudge_y = 0.005, nudge_x = 0)
gg3 <- gg2 + 
  theme(axis.text.y = element_blank())
#gg3

if (save_plots){
  ggsave2("{folder_fig}/07_{plotno}_{label}_ver1.png", gg1, 
         width = 8.5, height = 10.5, dpi = 500)
  # ggsave2("{folder_fig}/07_{plotno}_{label}_ver1.svg", gg1, 
  #        width = 8.5, height = 10.5, dpi = 500)
  ggsave2("{folder_fig}/07_{plotno}_{label}_ver2.png", gg2, 
         width = 8.5, height = 10.5, dpi = 500)
  # ggsave2("{folder_fig}/07_{plotno}_{label}_ver2.svg", gg2, 
  #        width = 8.5, height = 10.5, dpi = 500)
  ggsave2("{folder_fig}/07_{plotno}_{label}_ver3.png", gg3, 
         width = 8.5, height = 10.5, dpi = 500)
  # ggsave2("{folder_fig}/07_{plotno}_{label}_ver3.svg", gg3, 
  #        width = 8.5, height = 10.5, dpi = 500)
}

gg1
gg2
gg3


```
